@@ -7,8 +7,9 @@ import '../players/player_avatar.dart';
 class DisplayLeader extends StatelessWidget {
   final PlayerModel player;
   final Record record;
+  final int winningStreak;
 
-  DisplayLeader({@required this.player, @required this.record});
+  DisplayLeader({@required this.player, @required this.record, this.winningStreak});
 
   @override
   Widget build(BuildContext context) {
@@ -19,7 +20,9 @@ class DisplayLeader extends StatelessWidget {
               PlayerAvatar(player: player, radius: 50),
               Column(children: <Widget>[
                 Text(player.name, style: TextStyle(fontSize: 40)),
-                Text(record.additionalInfo(), style: TextStyle(fontSize: 18, color: Colors.black54))
+                if (record.additionalInfo() != '') Text(record.additionalInfo(), style: TextStyle(fontSize: 18, color: Colors.black54)),
+                if (winningStreak != null && winningStreak > 1)
+                  Text('x $winningStreak', style: TextStyle(fontSize: 18, color: Colors.black54))
               ])
             ], mainAxisAlignment: MainAxisAlignment.spaceAround)));
   }
